Add functionality for generating fake server responses
This commit introduces the functionality for generating fake server responses, ideal for testing, along with an associated server test. It adds a number of packages including 'pipe', 'logger' and 'server', and a main and cmd go file to execute commands. Additionally, it includes legal and descriptive files: a License and a Readme.